Add basic API and events (#311)
* Add API and events

* added proper api docs and api events

* added auto generated kotlin docs to page

+++ improved-factions/src/main/kotlin/io/github/toberocat/improvedfactions/factions/Faction.kt
@@ -17,6 +17,9 @@ import io.github.toberocat.improvedfactions.modules.power.PowerRaidsModule.power
 import io.github.toberocat.improvedfactions.modules.relations.RelationsModule
 import io.github.toberocat.improvedfactions.ranks.FactionRank
 import io.github.toberocat.improvedfactions.ranks.FactionRankHandler
+import io.github.toberocat.improvedfactions.api.events.FactionDeleteEvent
+import io.github.toberocat.improvedfactions.api.events.FactionJoinEvent
+import io.github.toberocat.improvedfactions.api.events.FactionLeaveEvent
 import io.github.toberocat.improvedfactions.ranks.listRanks
 import io.github.toberocat.improvedfactions.translation.LocalizationKey
 import io.github.toberocat.improvedfactions.translation.LocalizedException
@@ -104,6 +107,10 @@ class Faction(id: EntityID<Int>) : IntEntity(id) {
         }
 
     override fun delete() {
+        val event = FactionDeleteEvent(this)
+        Bukkit.getPluginManager().callEvent(event)
+        if (event.isCancelled()) return
+        
         BaseModule.claimChunkClusters.removeFactionClusters(this)
         DynmapModule.dynmapModule().dynmapModuleHandle.removeHome(this)
         listRanks().forEach { it.delete() }
@@ -112,7 +119,7 @@ class Faction(id: EntityID<Int>) : IntEntity(id) {
         }
         members().forEach { unsetUserData(it) }
         RelationsModule.deleteFactionRelations(id.value)
-
+        
         super.delete()
     }
 
@@ -182,6 +189,10 @@ class Faction(id: EntityID<Int>) : IntEntity(id) {
         if (isBanned(user))
             throw CommandException("base.exceptions.you-are-banned", emptyMap())
 
+        val event = FactionJoinEvent(this, user)
+        Bukkit.getPluginManager().callEvent(event)
+        if (event.isCancelled()) return
+        
         user.factionId = id.value
         user.assignedRank = rankId
 
@@ -319,8 +330,13 @@ class Faction(id: EntityID<Int>) : IntEntity(id) {
     @Throws(PlayerIsOwnerLeaveException::class)
     fun leave(player: UUID) {
         if (owner == player) throw PlayerIsOwnerLeaveException()
-
-        unsetUserData(player.factionUser())
+        
+        val user = player.factionUser()
+        val event = FactionLeaveEvent(this, user)
+        Bukkit.getPluginManager().callEvent(event)
+        if (event.isCancelled()) return
+        
+        unsetUserData(user)
         broadcast(
             "base.faction.player-left", mapOf(
                 "player" to (Bukkit.getOfflinePlayer(player).name ?: "unknown")
